# openapi.yaml
openapi: 3.0.1
info:
  title: 'SLS'
  version: 1.0.0
paths:
  '/users': 
    get:
    
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: phoneNumber
          in: query
          schema:
            type: string
            
      operationId: getUser
      
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                
      
    
    post:
      operationId: postUser
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - phoneNumber
              - fname
              - lname
              properties:
                phoneNumber:
                  type: string
                  description: phone number
                fname:
                  type: string
                  description: fname
                lname:
                  type: string
                  description: lname
                  
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
                
      

    

    patch:
      operationId: patchUser
      
      parameters:
      - name: userId
        required: true
        in: query
        schema:
          type: integer
      - name: phoneNumber
        required: true
        in: query
        schema:
          type: string
          
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/User'
                
          
      responses:
        '200':
          description: ok
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        
      
                
    
    delete:
      operationId: deleteUser
      
      parameters:
      - name: userId
        required: true
        in: query
        schema:
          type: integer
      - name: phoneNumber
        required: true
        in: query
        schema:
          type: string
      - name: userIds
        required: true
        in: query
        schema:
          type: array
          items:
            type: integer
        style: form
        explode: false
            
      responses: 
        '200':
          description: ok
      
    
  
components:
  schemas:
    User:
      title: user
      type: object
      required:
        - phoneNumber
        - fname
        - lname
        - userId
      description: ''
      properties:
        userId:
          type: integer
          description: userId
        phoneNumber:
          type: string
          description: phone number
        fname:
          type: string
          description: fname
        lname:
          type: string
          description: lname